change structure of files. add migration
